Update Spring example to use sentry-spring.
