+++ java/spring-boot/pom.xml
@@ -4,12 +4,11 @@
 	<modelVersion>4.0.0</modelVersion>
 
 	<groupId>io.sentry.example</groupId>
-	<artifactId>sentry-spring-example</artifactId>
+	<artifactId>sentry-spring-boot-example</artifactId>
 	<version>0.0.1</version>
 	<packaging>jar</packaging>
 
-	<name>demo</name>
-	<description>Demo project for Spring Boot</description>
+	<name>Sentry Spring-Boot Example</name>
 
 	<parent>
 		<groupId>org.springframework.boot</groupId>
@@ -27,8 +26,8 @@
 	<dependencies>
 		<dependency>
 			<groupId>io.sentry</groupId>
-			<artifactId>sentry-logback</artifactId>
-			<version>1.3.0</version>
+			<artifactId>sentry-spring</artifactId>
+			<version>1.3.1-SNAPSHOT</version>
 		</dependency>
 
 		<dependency>
@@ -56,6 +55,4 @@
 			</plugin>
 		</plugins>
 	</build>
-
-
 </project>
