--- java/spring/src/main/java/io/sentry/example/Application.java
@@ -1,37 +1,0 @@
-package io.sentry.example;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@Controller
-@EnableAutoConfiguration
-@SpringBootApplication
-public class Application {
-    private static final Logger logger = LoggerFactory.getLogger("example.Application");
-
-    @RequestMapping("/")
-    @ResponseBody
-    String home() {
-        logger.debug("Debug message");
-        logger.info("Info message");
-        logger.warn("Warn message");
-
-        try {
-            int example = 1 / 0;
-        } catch (Exception e) {
-            logger.error("Caught exception!", e);
-        }
-
-        return "Hello World!";
-    }
-
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-}
